+++ Sources/KeyboardKit/Appearance/StandardKeyboardAppearance.swift
@@ -15,25 +15,48 @@ import UIKit
  
  You can inherit this class then override any parts you like
  to customize the standard appearance.
+ 
+ `TODO` Unit test.
  */
 open class StandardKeyboardAppearance: KeyboardAppearance {
     
-    public init() {}
-    
-    open func font(for action: KeyboardAction) -> UIFont {
-        action.standardButtonFont
+    public init(context: KeyboardContext) {
+        self.context = context
     }
     
-    open func fontWeight(for action: KeyboardAction) -> UIFont.Weight? {
-        let hasImage = image(for: action) != nil
-        return hasImage ? .light : nil
+    private let context: KeyboardContext
+    
+    open func buttonBackgroundColor(for action: KeyboardAction) -> Color {
+        action.standardButtonBackgroundColor(for: context)
     }
     
-    public func image(for action: KeyboardAction) -> Image? {
+    open func buttonFont(for action: KeyboardAction) -> Font {
+        let rawFont = Font(action.standardButtonFont)
+        guard let weight = fontWeight(for: action) else { return rawFont }
+        return rawFont.weight(weight)
+    }
+    
+    open func buttonForegroundColor(for action: KeyboardAction) -> Color {
+        action.standardButtonForegroundColor(for: context)
+    }
+    
+    open func buttonImage(for action: KeyboardAction) -> Image? {
         action.standardButtonImage
     }
     
-    open func text(for action: KeyboardAction) -> String? {
+    open func buttonShadowColor(for action: KeyboardAction) -> Color {
+        action.standardButtonShadowColor(for: context)
+    }
+    
+    open func buttonText(for action: KeyboardAction) -> String? {
         action.standardButtonText
+    }
+}
+
+private extension StandardKeyboardAppearance {
+    
+    func fontWeight(for action: KeyboardAction) -> UIFont.Weight? {
+        let hasImage = buttonImage(for: action) != nil
+        return hasImage ? .light : nil
     }
 }