+++ RELEASE_NOTES.md
@@ -40,7 +40,6 @@ Although these release notes will aim at covering everything that changes in thi
 * `KeyboardAction.emoji` now uses the `Emoji` type instead of a char.
 * `KeyboardAction.emojiCategory` no longer has a  standard tap action.
 * `KeyboardActionRow` `standardButtonImage` no longer takes a context.
-* `KeyboardAppearance` `fontWeight` no longer takes a context.
 * `KeyboardContext` no longer has any services, just inspectable properties.
 * `KeyboardInputViewController` `context` is now an `ObservableKeyboardContext`.
 * `KeyboardInputViewController` `context` is now called `keyboardContext`.
@@ -55,6 +54,8 @@ Although these release notes will aim at covering everything that changes in thi
 * All previews deprecations have been removed in this version.
 * `InputCalloutContext.shared` has been removed. Use the environment object instead.
 * `KeyboardActionRow` has been removed, since it's confusing to have two aliases for the same thing.
+* `KeyboardAppearance` button properties are prefixed with `button`.
+* `KeyboardAppearance` `font` and `fontWeight` have been merged into a single `font` property.
 * `KeyboardAppearanceProvider` has been renamed to `KeyboardAppearance`.
 * `KeyboardContext` `actionHandler` has been moved to the input vc.
 * `KeyboardContext` `keyboardAppearanceProvider` has been moved to the input vc.